Make Http client import optional
Refactor the so now the primary export of the module doesn't have the
HTTP client attached to it, and now its only used by the server/cli
functions, and consumers have to bring their own `fetch` function.
